fix: decouple desktop diagnostics timeout
