+++ desktop/scripts/packaged-diagnostics.mjs
@@ -1,3 +1,7 @@
+import { readFile, stat } from "node:fs/promises";
+
+const DIAGNOSTIC_JOURNAL_TIMEOUT_MS = 5_000;
+
 const processGoneReasons = new Set([
   "clean-exit",
   "abnormal-exit",
@@ -102,4 +106,71 @@ export function verifyPackagedDiagnosticEvent(event) {
   throw new Error(
     "packaged diagnostic journal contains an unexpected startup event",
   );
+}
+
+export async function verifyPackagedDiagnosticJournal(
+  path,
+  timeoutMs = DIAGNOSTIC_JOURNAL_TIMEOUT_MS,
+) {
+  const deadline = Date.now() + timeoutMs;
+  while (Date.now() < deadline) {
+    try {
+      const body = await readFile(path, "utf8");
+      if (Buffer.byteLength(body, "utf8") > 128 * 1024) {
+        throw new Error("packaged diagnostic journal exceeds its size limit");
+      }
+      const document = JSON.parse(body);
+      if (
+        Object.keys(document).sort().join(",") !== "events,schemaVersion" ||
+        document.schemaVersion !== 1 ||
+        !Array.isArray(document.events) ||
+        document.events.length === 0 ||
+        document.events.length > 256
+      ) {
+        throw new Error(
+          "packaged diagnostic journal has an unexpected manifest",
+        );
+      }
+      if (
+        !document.events.some(
+          (event) => event?.kind === "startup" && event.packaged === true,
+        )
+      ) {
+        throw new Error(
+          "packaged diagnostic journal is missing its startup event",
+        );
+      }
+      for (const event of document.events) {
+        verifyPackagedDiagnosticEvent(event);
+      }
+      if (
+        /https?:|origin|cookie|token|authorization|console|filename/iu.test(
+          body,
+        )
+      ) {
+        throw new Error(
+          "packaged diagnostic journal contains forbidden sensitive fields",
+        );
+      }
+      if (
+        process.platform !== "win32" &&
+        ((await stat(path)).mode & 0o077) !== 0
+      ) {
+        throw new Error(
+          "packaged diagnostic journal permissions are not private",
+        );
+      }
+      return;
+    } catch (error) {
+      if (
+        error instanceof SyntaxError ||
+        (error instanceof Error &&
+          !("code" in error && error.code === "ENOENT"))
+      ) {
+        throw error;
+      }
+    }
+    await new Promise((resolveDelay) => setTimeout(resolveDelay, 50));
+  }
+  throw new Error("packaged application did not persist diagnostics");
 }
